[RFC] Buffer Layout Padding
This RFC introduces a method to specify padding to be applied as part
of a buffer layout transformation, to be used when the desired layout
does not evenly tile the buffer being transformed, and simplifications
that can be performed based on these padded buffers.

The motivating examples are primarily in the "Implementation options"
section, which goes through several desired usages of the buffer
padding, and how they can be automatically derived using the
TIR primitives/transformations described in earlier sections.