Make it clear that tag is enforced but capitalization is not
Original proposal for this RFC had tag presence as an enforced rule.
Later, text on recommendation about tag capitalization was added to the
rule and it became a tad unclear if the enforcement applies to the tag
presence only or also to the tag capitalization text.

This commit moves the "enforced" mark before the tag capitalization text
to make it clearer that it applies to the tag presence only. It also
tweaks a bit the text to avoid the idea that tag presence is optional --
since it's enforced, just like it was in the original proposal.