[RFC] Add Commit Message Guideline (#88)
* [RFC] Add Commit Message Guideline

This RFC proposes adding a Commmit Message Guideline to Apache TVM
documentation to help guide contributors on how to write good commit
messages when submitting code / Pull Requests.

Co-authored-by: Leandro Nunes <leandro.nunes@arm.com>

* Remove out-of-scope paragraph

As per the reviews, remove this paragraph since multiple commits / changes in a
PR is out-of-scope, best fits in another guideline, and would require
addtional discussions, needing a separate vote.

* Be more explicit about committer and author responsibilities

Be more explicit about committer and author responsibilities when
the PR title and/or body need to get updated to be in sync with the code
changes after a review cycle.

* Clarify about capitalization in tags

Clarify about the preferred capitalization in tags. Also clarify about
multiple tags and acronyms when used as a tag.

* Be more emphatic about no need of a period at the end of the title

Be more emphatic about no need of a period at the end of the title but
don't enforce it.

* Make it clear that tag is enforced but capitalization is not

Original proposal for this RFC had tag presence as an enforced rule.
Later, text on recommendation about tag capitalization was added to the
rule and it became a tad unclear if the enforcement applies to the tag
presence only or also to the tag capitalization text.

This commit moves the "enforced" mark before the tag capitalization text
to make it clearer that it applies to the tag presence only. It also
tweaks a bit the text to avoid the idea that tag presence is optional --
since it's enforced, just like it was in the original proposal.

* Update rfcs/0088-commit-message-guideline.md

Co-authored-by: Siyuan Feng <Hzfengsy@sjtu.edu.cn>

Co-authored-by: Leandro Nunes <leandro.nunes@arm.com>
Co-authored-by: Siyuan Feng <Hzfengsy@sjtu.edu.cn>